fixed a bunch

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: lpraca-l <lplacerdadesign@gmail.com>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/08 16:19:41 by lpraca-l      #+#    #+#                 */
-/*   Updated: 2022/10/09 21:52:16 by lpraca-l      ########   odam.nl         */
+/*   Updated: 2022/10/21 16:39:20 by lpraca-l      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@ The strlen() function returns the number of bytes in the string pointed to by s.
 
 #include "libft.h"
 
-int	ft_strlen(const char *c)
+size_t	ft_strlen(const char *c)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (c[count] != '\0')
