+++ libft/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: lpraca-l <lpraca-l@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/17 14:14:55 by lpraca-l      #+#    #+#                 */
-/*   Updated: 2022/10/21 13:51:44 by lpraca-l      ########   odam.nl         */
+/*   Updated: 2022/10/21 16:40:31 by lpraca-l      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	sub = malloc((len +1) * sizeof(char));
 	if (sub == NULL)
 		return (NULL);
-	if (ft_strlen(s) < (int)start)
+	if (ft_strlen(s) < start)
 		len = 0;
 	while (s[start] != '\0' && count < len)
 	{
