+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: lpraca-l <lplacerdadesign@gmail.com>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/09 19:47:08 by lpraca-l      #+#    #+#                 */
-/*   Updated: 2022/10/21 09:22:14 by lpraca-l      ########   odam.nl         */
+/*   Updated: 2022/10/21 16:39:42 by lpraca-l      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ int		ft_isprint(int c);
 void	*ft_memcpy(void *dest, const void *src, size_t n);
 void	*ft_memmove(void *dest, const void *src, size_t n);
 void	*ft_memset(void *s, int c, size_t n);
-int		ft_strlen(const char *c);
+size_t	ft_strlen(const char *c);
 size_t	ft_strlcpy(char	*dst, const char	*src, size_t dstsize);
 int		ft_toupper(int c);
 int		ft_tolower(int c);
